requirements.txt
biopython==1.73
progressbar2==3.39.3
nose==1.3.7
mock==3.0.5
flake8==3.7.7
ftputil==3.4
click==7.1.2
tqdm~=4.58.0
setuptools~=52.0.0